requirements.txt
# まえのやつ
gitpython
ipython
matplotlib
numpy==1.26.4
opencv-python
Pillow
psutil
PyYAML
requests
scipy
thop
torch
torchvision
tqdm
torchaudio
tensorboard
pandas
seaborn
albumentations
pycocotools
Flask
pyserial
ipykernel
flask-cors
torch==2.3.1+cu121
torchaudio==2.3.1+cu121
torchvision==0.18.1+cu121
Flask-SocketIO==5.4.1

# GPUなし用(町田さん)
# absl-py==2.1.0
# albucore==0.0.17
# albumentations==1.4.17
# annotated-types==0.7.0
# argostranslate==1.9.6
# asttokens==2.4.1
# bidict==0.23.1
# blinker==1.8.2
# certifi==2024.8.30
# charset-normalizer==3.3.2
# click==8.1.7
# colorama==0.4.6
# comm==0.2.2
# contourpy==1.3.0
# ctranslate2==4.4.0
# cycler==0.12.1
# debugpy==1.8.1
# decorator==5.1.1
# eval_type_backport==0.2.0
# exceptiongroup==1.2.2
# executing==2.1.0
# filelock==3.16.1
# Flask==3.0.3
# Flask-SocketIO==5.4.1
# fonttools==4.54.1
# fsspec==2024.9.0
# gitdb==4.0.11
# GitPython==3.1.43
# grpcio==1.66.2
# h11==0.14.0
# idna==3.10
# imageio==2.35.1
# ipykernel==6.29.4
# ipython==8.28.0
# itsdangerous==2.2.0
# jedi==0.19.1
# Jinja2==3.1.4
# joblib==1.4.2
# jupyter_client==8.6.2
# jupyter_core==5.7.2
# keyboard==0.13.5
# kiwisolver==1.4.7
# labelme==3.16.7
# lazy_loader==0.4
# Markdown==3.7
# MarkupSafe==2.1.5
# matplotlib==3.9.2
# matplotlib-inline==0.1.7
# mpmath==1.3.0
# nest-asyncio==1.6.0
# networkx==3.3
# numpy==2.1.2
# opencv-python==4.10.0.84
# opencv-python-headless==4.10.0.84
# packaging==24.1
# pandas==2.2.3
# parso==0.8.4
# pillow==10.4.0
# platformdirs==4.2.2
# prompt_toolkit==3.0.48
# protobuf==5.28.2
# psutil==6.0.0
# pure_eval==0.2.3
# pycocotools==2.0.8
# pydantic==2.9.2
# pydantic_core==2.23.4
# Pygments==2.18.0
# pyparsing==3.1.4
# pyperclip==1.8.2
# PyQt5==5.15.11
# PyQt5-Qt5==5.15.2
# PyQt5_sip==12.15.0
# pyserial==3.5
# PySide6==6.4.1
# PySide6-Addons==6.4.1
# PySide6-Essentials==6.4.1
# python-dateutil==2.9.0.post0
# python-engineio==4.9.1
# python-socketio==5.11.4
# pytz==2024.2
# pywin32==306
# PyYAML==6.0.2
# pyzmq==26.0.3
# QtPy==2.4.1
# regex==2024.9.11
# requests==2.32.3
# sacremoses==0.0.53
# scikit-image==0.24.0
# scipy==1.14.1
# seaborn==0.13.2
# sentencepiece==0.2.0
# shiboken6==6.4.1
# simple-websocket==1.1.0
# six==1.16.0
# smmap==5.0.1
# stack-data==0.6.3
# stanza==1.1.1
# sympy==1.13.3
# tensorboard==2.18.0
# tensorboard-data-server==0.7.2
# termcolor==2.4.0
# thop==0.1.1.post2209072238
# tifffile==2024.9.20
# torch==2.4.1
# torchaudio==2.4.1
# torchvision==0.19.1
# tornado==6.4
# tqdm==4.66.5
# traitlets==5.14.3
# typing_extensions==4.12.2
# tzdata==2024.2
# urllib3==2.2.3
# wcwidth==0.2.13
# Werkzeug==3.0.4
# wsproto==1.2.0